opioid results.sh for reproducibility of different RXCUI cohort selections.

+++ opioid/results.sh
@@ -8,12 +8,20 @@ $mysql_dataset -e "call log('refresh', 'begin')"
 $mysql_dataset < drop_tables.sql
 $mysql_dataset < create_tables.sql
 
-#export CURATED="VSAC_Mathematica"
-#export CURATED="curated_from_bioportal"
-#export CURATED="expanded_from_bioportal"
-export CURATED="curated_from_expanded_march26"
+# export CURATED="all_rxcui_str"
+# export CURATED="VSAC_Mathematica"
+# export CURATED="bioportal"
+# export CURATED="bioportal_to_umls"
+# export CURATED="umls_reviewed_march26"
+# export CURATED="curated_from_expanded_april7"
+
+if [ -z "$CURATED" ]; then
+  echo "CURATED is not set. Exiting..."
+  exit 1
+fi
 
 echo "Using $CURATED.tsv"
+$mysql_dataset -e "call log('$CURATED', ''${CURATED}'')"
 $mysql_dataset -e "call log('${CURATED}', 'begin')"
 
 rm -f curated.tsv
@@ -25,8 +33,10 @@ $mysql_dataset < stats.sql
 pushd .
 cd ..
 ./backup_curated.sh opioid curated
-./backup_curated.sh opioid curated
+./backup_curated.sh opioid expand
 
+./export_tsv.sh opioid stats_expand
+./export_tsv.sh opioid stats_keywords
 ./export_tsv.sh opioid stats_sab
 ./export_tsv.sh opioid stats_tty
 ./export_tsv.sh opioid stats_tui
@@ -36,16 +46,3 @@ cd ..
 cd opioid
 
 $mysql_dataset -e "call log('${CURATED}', 'done')"
-
-# ./export_bsv.sh rxnorm expand
-# ./export_bsv.sh rxnorm expand_consists
-# ./export_bsv.sh rxnorm expand_doseform
-# ./export_bsv.sh rxnorm expand_form
-# ./export_bsv.sh rxnorm expand_ingredient
-# ./export_bsv.sh rxnorm expand_ingredients
-# ./export_bsv.sh rxnorm expand_isa
-# ./export_bsv.sh rxnorm expand_tradename
-# ./export_bsv.sh rxnorm expand_cui
-# ./export_bsv.sh rxnorm expand_cui_str
-# ./export_bsv.sh rxnorm expand_cui_rela_cui
-# ./export_bsv.sh rxnorm expand_cui_rela_str
